Highlight tmux active pane
Having the borders as the 'only' active pane indicator is pretty bad,
specially for people who are not used to my config, or for people who
are just spectating what I'm doing on the screen.

Changing the bg color of the active pane makes it easy to follow the
action.